+++ src/msdb5/card/card_err_test.go
@@ -9,7 +9,7 @@ func TestId41IsInvalid(t *testing.T) {
 	verifyInvalidID(t, 41)
 }
 
-func verifyInvalidID(t *testing.T, id int) {
+func verifyInvalidID(t *testing.T, id uint8) {
 	_, err := ByID(id)
 	if err == nil {
 		t.Fatalf("%d is not valid id", id)
